sửa lỗi

+++ Dt/past/de40/NT.cpp
@@ -3,7 +3,7 @@
 
 #define ull unsigned long long
 
-bool isPrime(int n) {
+bool isPrime(ull n) {
     if (n <= 3) return n > 1;
     if (n % 2 == 0 || n % 3 == 0) return false;
     for (int i = 5; i * i <= n; i += 6)
